organized components into seperate files and
