+++ anecdotes/src/components/Anecdote.js
@@ -1,0 +1,12 @@
+import React from "react"
+import Title from "./Title"
+import AnecdoteBody from "./AnecdoteBody"
+
+const Anecdote = ({ element, title, anecdote, voteScore }) => (
+    <>
+      <Title element={element} title={title} />
+      <AnecdoteBody anecdote={anecdote} voteScore={voteScore} />
+    </>
+  )
+
+  export default Anecdote
