+++ anecdotes/src/components/AnecdoteText.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const AnecdoteText = ({ anecdote }) => (
+    <span>{anecdote}</span>
+)
+
+export default AnecdoteText
