+++ anecdotes/src/components/Button.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Button = ({ children, onClick }) => (
+    <button onClick={onClick}>
+      {children}
+    </button>
+)
+
+export default Button
